code/CDNA/requirements.txt
absl-py                            0.11.0
alabaster                          0.7.10
appnope                            0.1.0
asn1crypto                         0.24.0
astor                              0.8.1
awscli                             1.11.104
Babel                              2.4.0
backports-abc                      0.5
backports.functools-lru-cache      1.6.1
backports.shutil-get-terminal-size 1.0.0
backports.weakref                  1.0.post1
bleach                             2.0.0
botocore                           1.5.67
bzr                                2.8.0.dev1
cachetools                         3.1.1
catkin-pkg                         0.4.23
catkin-pkg-modules                 0.4.23
catkin-tools                       0.6.1
certifi                            2017.4.17
chainer                            6.7.0
chardet                            3.0.4
click                              6.7
cloudpickle                        1.3.0
colorama                           0.3.7
configobj                          5.0.6
configparser                       3.5.0
coverage                           4.4.1
cryptography                       2.1.4
cupy-cuda100                       6.7.0
cycler                             0.10.0
decorator                          4.0.11
docutils                           0.13.1
entrypoints                        0.2.3
enum34                             1.1.6
fastrlock                          0.5
ffmpeg-python                      0.2.0
filelock                           2.0.11
flake8                             3.3.0
funcsigs                           1.0.2
functools32                        3.2.3.post2
future                             0.18.2
futures                            3.1.1
gast                               0.2.2
google-auth                        1.22.1
google-auth-oauthlib               0.4.1
google-pasta                       0.2.0
grpcio                             1.33.2
gyp                                0.1
h5py                               2.10.0
html5lib                           0.999999999
httplib2                           0.9.2
idna                               2.5
imageio                            2.2.0
imagesize                          0.7.1
ipaddress                          1.0.17
ipykernel                          4.6.1
ipython                            5.4.1
ipython-genutils                   0.2.0
ipywidgets                         6.0.0
Jinja2                             2.9.6
jmespath                           0.9.3
jsonschema                         2.6.0
jupyter                            1.0.0
jupyter-client                     5.0.1
jupyter-console                    5.1.0
jupyter-core                       4.3.0
Keras                              2.4.3
Keras-Applications                 1.0.8
Keras-Preprocessing                1.1.2
keyring                            10.6.0
keyrings.alt                       3.0
kiwisolver                         1.1.0
launchpadlib                       1.10.6
lazr.restfulclient                 0.13.5
lazr.uri                           1.0.3
Markdown                           3.1.1
MarkupSafe                         1.0
matplotlib                         2.0.2
mccabe                             0.6.1
mercurial                          4.5.3
mistune                            0.7.4
mock                               2.0.0
moviepy                            1.0.3
natsort                            6.2.1
nbconvert                          5.2.1
nbformat                           4.3.0
networkx                           2.2
nose                               1.3.7
notebook                           5.0.0
numpy                              1.16.0
oauth                              1.0.1
oauthlib                           3.1.0
olefile                            0.44
opencv-python                      4.2.0.32
opt-einsum                         2.3.2
osrf-pycommon                      0.1.9
pandas                             0.24.2
pandocfilters                      1.4.1
paramiko                           2.0.0
pathlib                            1.0.1
pathlib2                           2.3.0
pbr                                3.0.1
pexpect                            4.2.1
pickleshare                        0.7.4
Pillow                             4.1.1
pip                                20.3.3
proglog                            0.1.9
prompt-toolkit                     1.0.14
protobuf                           3.3.0
psutil                             5.4.2
ptyprocess                         0.5.1
pyasn1                             0.2.3
pyasn1-modules                     0.2.8
pycairo                            1.16.2
pycodestyle                        2.3.1
pycrypto                           2.6.1
pyflakes                           1.5.0
Pygments                           2.2.0
pygobject                          3.26.1
pyOpenSSL                          17.5.0
pyparsing                          2.2.0
python-dateutil                    2.6.0
python-dotenv                      0.6.4
pytz                               2017.2
PyWavelets                         1.0.3
pyxdg                              0.25
PyYAML                             3.12
pyzmq                              16.0.2
qtconsole                          4.3.0
requests                           2.17.3
requests-oauthlib                  1.3.0
roman                              2.0.0
rosdep                             0.20.0
rosdep-modules                     0.20.0
rosdistro                          0.8.3
rosdistro-modules                  0.8.3
rosinstall                         0.7.8
rosinstall-generator               0.1.22
rospkg                             1.2.9
rospkg-modules                     1.2.9
rsa                                3.4.2
s3transfer                         0.1.10
scandir                            1.5
scikit-image                       0.14.5
scikit-learn                       0.20.4
scipy                              1.2.2
SecretStorage                      2.3.1
setuptools                         44.1.1
simplegeneric                      0.8.1
simplejson                         3.13.2
singledispatch                     3.4.0.3
six                                1.10.0
snowballstemmer                    1.2.1
Sphinx                             1.6.2
sphinxcontrib-websupport           1.0.1
subprocess32                       3.2.7
tensorboard                        1.15.0
tensorflow                         1.1.0
tensorflow-estimator               1.15.1
tensorflow-gpu                     1.15.0
termcolor                          1.1.0
terminado                          0.6
testpath                           0.3.1
tornado                            4.5.1
tqdm                               4.50.2
traitlets                          4.3.2
trollius                           2.0.1
typing                             3.6.6
typing-extensions                  3.6.6
urllib3                            1.21.1
vcstools                           0.1.42
wadllib                            1.3.2
wcwidth                            0.1.7
webencodings                       0.5.1
Werkzeug                           0.12.2
wheel                              0.35.1
widgetsnbextension                 2.0.0
wrapt                              1.12.1
wstool                             0.1.17
zope.interface                     4.3.2
